bot/requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
pillow==11.0.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
pytubefix==6.13.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
